feat(bookmarks): add bookmark feature for nodes
Uses the signcolum to display the mark status.
Marks are saved in memory for now, we'll see if we want to implement a
filesystem save for the marks (would not be hard).
Added keybindings:
- ma -> add a mark on the node
- md -> remove a mark from the node
- M -> toggle the mark on the node
Users can access the marks through `_G.NvimTreeMarks` which is a record
of `{ [absolute_path] = true }`.
They can call `require "nvim-tree.marks".get_marks()` which returns a
list of absolute paths.